backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# LangChain & LangGraph for Multi-Agent Orchestration
langchain>=0.2.0
langgraph>=0.2.0
langchain-groq>=0.1.9
langchain-openai>=0.1.20
langchain-community>=0.2.0

# Supabase & Vector Storage
supabase>=2.4.0
pgvector>=0.2.5

# Auth
pyjwt[crypto]>=2.8.0
httpx>=0.27.0

# File Processing
pypdf>=4.0.1
python-docx>=1.1.0
tiktoken>=0.6.0

# Utilities
pydantic>=2.6.1
pydantic-settings>=2.1.0
aiofiles>=23.2.1

